added middleware to secure routes

+++ app/api/issues/[id]/route.ts
@@ -1,10 +1,14 @@
+import authOption from "@/app/auth/authOption";
 import { issueSchema } from "@/app/validationSchemas";
 import { prisma } from "@/prisma/client";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 export async function PATCH(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const server = await getServerSession(authOption);
+  if (!server) return NextResponse.json({}, { status: 401 });
   const body = await request.json();
   const validation = issueSchema.safeParse(body);
   if (!validation.success)
@@ -23,21 +27,23 @@ export async function PATCH(
     },
   });
 
-  return NextResponse.json(updateIssue)
+  return NextResponse.json(updateIssue);
 }
 
 export async function DELETE(
   request: NextRequest,
   { params }: { params: { id: string } }
 ) {
+  const server = await getServerSession(authOption);
+  if (!server) return NextResponse.json({}, { status: 401 });
   const issue = await prisma.issue.findUnique({
-    where:{id:parseInt(params.id)}
-  })
-  if(!issue)
-    return NextResponse.json({error:'Issue Not Found'},{status:404})
+    where: { id: parseInt(params.id) },
+  });
+  if (!issue)
+    return NextResponse.json({ error: "Issue Not Found" }, { status: 404 });
   await prisma.issue.delete({
-    where:{id:issue?.id}
-  })
+    where: { id: issue?.id },
+  });
 
-  return NextResponse.json({})
+  return NextResponse.json({});
 }
